Refactor to function [1/2]
I'm not thrilled with this, but it works. I had to ditch the original SlidingTabLayout and Collapsing Toolbar for functionality.

This is based around PagerSlidingTabStrip from:
https://github.com/astuetz/PagerSlidingTabStrip

Also, renamed the packages to be more uniformed. Like I said, I'm not thrilled with this layout as I wanted to be with a different look, and it might resemble other projects' work, but it beats tapping through a bunch of preference screens.

Major thanks to @Mazda-- for pointing out my original mistakes with my actiity and pointing me in the right direction.

PS2:
Fixup PowerMenuFragment to launch correctly

Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>