Allow Heads up and Ticker to play nice once and for all
Until now users could enable both ticker and heads up and while only one
would work at a time, we would get reports of one not working. This was
common sense for most but not for all.

This commit takes care of that issue and protects the user from the user.
You can now only enable one at a time. So if heads up is enabled the ticker
is disabled and if the ticker was enabled, heads is disabled.

- Renamed ticker preference
- Moved ticker under Mulitasking

Change-Id: Ic7b9a509d954b0b0aee331759f85fcf45fb457ea
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>